requirements.txt
antlr4-python3-runtime==4.10  # Runtime di ANTLR per Python, per il supporto del parser
